Fix na podstawie zgłoszenia w panelu
1. resolv.conf not working after install resolvconf
2. Go is not visible for wireguard-go (needed for compile) - linked to /usr/bin/go
3. Typo: devices -> device

+++ scripts/chce_wireguard.sh
@@ -2,11 +2,23 @@
 # Autor skryptu: Andrzej Szczepaniak
 # Poprawki: Jakub 'unknow' Mrugalski
 
+cp /etc/resolv.conf /etc/resolv.back
+
 apt update
 apt install -y --no-install-recommends libmnl-dev make qrencode wireguard-tools resolvconf
 
+systemctl stop resolvconf
+systemctl disable resolvconf
+mv /etc/resolv.back /etc/resolv.conf
+
+if [ ! -e /dev/net/tun ]; then
+    echo "Aby Wireguard działał poprawnie, musisz aktywować TUN/TAP na swoim serwerze";
+    exit
+fi
+
 cd /tmp/ && wget https://git.zx2c4.com/wireguard-go/snapshot/wireguard-go-0.0.20191012.tar.xz && tar -xvf wireguard-go-0.0.20191012.tar.xz && rm wireguard-go-0.0.20191012.tar.xz && mv wireguard-go-0.0.20191012 wireguard-go
 git -c http.sslVerify=false clone https://git.zx2c4.com/wireguard-tools /tmp/wireguard-tools
+
 
 wget https://dl.google.com/go/go1.16.4.linux-amd64.tar.gz -O /tmp/go1.16.4.linux-amd64.tar.gz
 
@@ -15,14 +27,16 @@ mv /tmp/go /usr/local
 export GOROOT=/usr/local/go
 export PATH=$GOPATH/bin:$GOROOT/bin:$PATH
 
-cd /tmp/wireguard-tools/src/
+ln -s /usr/local/go/bin/go /usr/bin/
+
+cd /tmp/wireguard-tools/src/ || exit
 make
 make install
 
-cd /tmp/wireguard-go/devices/
-rm /tmp/wireguard-go/devices/queueconstants_default.go && wget https://fx.vc-mp.eu/shared/queueconstants_default.go
+cd /tmp/wireguard-go/device/
+rm /tmp/wireguard-go/device/queueconstants_default.go && wget https://fx.vc-mp.eu/shared/queueconstants_default.go
 cd /tmp/wireguard-go && make
-cp wireguard-go /usr/local/bin && ln -s /usr/local/go/bin/go /usr/bin/
+cp /tmp/wireguard-go/wireguard-go /usr/bin
 
 wg genkey > /etc/wireguard/privatekey
 wg genkey > /etc/wireguard/client-privatekey
@@ -31,7 +45,7 @@ chmod 600 /etc/wireguard/*privatekey*
 
 cat /etc/wireguard/privatekey | wg pubkey > /etc/wireguard/publickey
 cat /etc/wireguard/client-privatekey | wg pubkey > /etc/wireguard/client-publickey
-srv=`hostname`
+srv=$(hostname)
 privkey=$(cat /etc/wireguard/privatekey)
 pubkey=$(cat /etc/wireguard/publickey)
 cliprivkey=$(cat /etc/wireguard/client-privatekey)
